% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assignVfCriteria}
\alias{assignVfCriteria}
\title{VF criteria application}
\usage{
assignVfCriteria()
}
\description{
Apply Foster, MHPA, UKGTS criteria to all patients from a properly-formatted excel spreadsheet
}
