% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkFostCriteria}
\alias{checkFostCriteria}
\title{Foster criteria (FOST)}
\usage{
checkFostCriteria(pat_id)
}
\arguments{
\item{pat_id}{Patient position in the list on the excel spreadsheet}
}
\value{
BOOL Whether patient VF analysis resulted in a positive result.
}
\description{
Glaucoma Hemifield Test (GHT) “outside normal limits” AND a cluster of three contiguous points at the 5% level on the pattern deviation plot
}
