% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkGhtCriteria}
\alias{checkGhtCriteria}
\title{GHT criteria (GHT)}
\usage{
checkGhtCriteria(pat_id)
}
\arguments{
\item{pat_id}{Patient position in the list on the excel spreadsheet}
}
\value{
BOOL Whether patient VF analysis resulted in a positive result.
}
\description{
Glaucoma Hemifield Test (GHT) “outside normal limits”
}
