% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkMhpaCriteria}
\alias{checkMhpaCriteria}
\title{Modified Hoddap-Parrish-Anderson criteria (MHPA)}
\usage{
checkMhpaCriteria(pat_id)
}
\arguments{
\item{pat_id}{Patient position in the list on the excel spreadsheet}
}
\value{
BOOL Whether patient VF analysis resulted in a positive result.
}
\description{
GHT outside normal limits; OR cluster of 3 points on the pattern deviation plot depressed at p<5%, one of which depressed at p<1%; OR PSD with p<5%
}
