% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkUkgtsCriteria}
\alias{checkUkgtsCriteria}
\title{UKGTS criteria}
\usage{
checkUkgtsCriteria(pat_id)
}
\arguments{
\item{pat_id}{Patient position in the list on the excel spreadsheet}
}
\value{
BOOL Whether patient VF analysis resulted in a positive result.
}
\description{
2 or more contiguous points with P<0.01 loss or more; OR 3 or more contiguous points with P<0.05 loss or more; OR a 10-dB difference across the nasal horizontal midline at 2 or more adjacent points in the total deviation plot.
}
