% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printClusteredHist}
\alias{printClusteredHist}
\title{Clustered histogram for MD and criteria results}
\usage{
printClusteredHist(percentage.weight = FALSE)
}
\description{
Print histogram where criteria (Foster, MHPA, UKGTS, and GHT) results are clustered within their MD percentiles (10, 5, 2, 1, and 0.5 %)
}
