% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printEulerDiag}
\alias{printEulerDiag}
\title{Euler diagram}
\usage{
printEulerDiag(df.results = df.criteria_results)
}
\description{
Print euler diagram (i.e. weighted venn diagram) for the overlapping criteria results groups
}
