% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printIntersectionBarGraph}
\alias{printIntersectionBarGraph}
\title{Intersetion bar graph}
\usage{
printIntersectionBarGraph(df.results = df.criteria_results)
}
\description{
Print intersection bar graph to compare overlapping criteria results groups
}
