% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printVfMap}
\alias{printVfMap}
\title{Partial STATPAC SFA}
\usage{
printVfMap(pat_id)
}
\arguments{
\item{pat_id}{Patient position in the list on the excel spreadsheet}
}
\value{
grid_plot The grid plot of the STATPAC SFA analysis.
}
\description{
Print reliability and global indices, criteria results, and VF maps (TD, TDP, PD, PDP) for a given patient
}
